### 2017-07-11 #317 @whedon
[PRE REVIEW]: mcMST: A Toolbox for the Multi-Criteria Minimum Spanning Tree Problem
**Submitting author:** @jakobbossek (<a href="http://orcid.org/0000-0002-4121-4668">Jakob Bossek</a>)
**Repository:** <a href="https://github.com/jakobbossek/mcMST" target ="_blank">https://github.com/jakobbossek/mcMST</a>
**Version:** v1.0.0
