### 2017-08-15 #317 @jakobbossek
@arfon Thanks for having a first look at my submission!

> Could you confirm here that there is a research application for this software (and explain what that
> application is)? The section 'what should my paper contain' has some guidance for the sort of
> content we're looking to be present in the paper.md.

The software is beneficial for researchers who work in the field of evolutionary multi-objecitve optimization of combinatorial (graph) problems and the multi-criteria spanning tree problem (mcMST) in particular. An example: Most recently we used the software to develop a new mutation operator for an evolutionary multi-objective algorithm for the mcMST based on insights gained from the empirical analysis of Pareto-optimal solutions of mcMST problems. This work was submitted to the multi-criteria decision making track of the [IEEE SSCI2017](http://www.ele.uri.edu/ieee-ssci2017/) and is currently under review.
