### 2017-08-19 #317 @arfon
OK thanks @jakobbossek.

> An example: Most recently we used the software to develop a new mutation operator for an evolutionary multi-objective algorithm for the mcMST based on insights gained from the empirical analysis of Pareto-optimal solutions of mcMST problems. This work was submitted to the multi-criteria decision making track of the IEEE SSCI2017 and is currently under review.

If you haven't already, it might be worth adding something about this to the `paper.md`.

We just need to find an editor and reviewer for this submission now before we can proceed with the main review.
