### 2017-08-21 #317 @Kevin-Mattheus-Moerman
@ArdiaD @lmarti @berndbischl @jakob-r @aklxao2 @luca-scr @gvegayon would you be interested in reviewing this submission for [JOSS](http://joss.theoj.org/)? If you choose to accept this invitation we'll open a review issue where the actual review will take place. [Review](http://joss.theoj.org/about#reviewer_guidelines) is often fast and focuses on the software, its claims, and the documentation. Thanks! :rocket:
