### 2017-08-21 #317 @gvegayon
Hi Kevin,

I'm in. Thanks for inviting me.

Best,

George G. Vega Yon (from mobile)
+1 626 381 8171
http://cana.usc.edu/vegayon

On Aug 21, 2017 07:46, "Kevin Mattheus Moerman" <notifications@github.com>
wrote:

@ArdiaD <https://github.com/ardiad> @lmarti <https://github.com/lmarti>
@berndbischl <https://github.com/berndbischl> @jakob-r
<https://github.com/jakob-r> @aklxao2 <https://github.com/aklxao2> @luca-scr
<https://github.com/luca-scr> @gvegayon <https://github.com/gvegayon> would
you be interested in reviewing this submission for JOSS
<http://joss.theoj.org/>? If you choose to accept this invitation we'll
open a review issue where the actual review will take place. Review
<http://joss.theoj.org/about#reviewer_guidelines> is often fast and focuses
on the software, its claims, and the documentation. Thanks! 🚀
