### 2017-08-21 #374 @Kevin-Mattheus-Moerman
:rocket: @jakobbossek This is where the review process will happen.

@gvegayon see here for the [reviewer guidelines](http://joss.theoj.org/). See here for an example [review process](https://github.com/openjournals/joss-reviews/issues/354). Feel free to comment here or create issues on the repository and link to them here. Tick the boxes at the top here if you feel the requirements are met. Let me know if you have questions.
