### 2017-08-21 #374 @berndbischl
hi guys,

i am
a) currently traveling
and
b) have worked together with jakob on multiple suggestions, including suoervising his master's, so i dont feel there is an arm's length distance here

will therefore decline.

(PS: I want to say how much I like everything I see what happens at JOSS)
