### 2017-08-22 jakobbossek/mcMST#8 @gvegayon
The AppVeyor link is broken
The link on the AppVeyor badge is broken https://ci.appveyor.com/project/jakobbossek/mcMST/branch/master
