### 2017-08-22 jakobbossek/mcMST#9 @gvegayon
The @return tag in mcMSTPrim.R is bad
```tex
[list] List with component pareto.front. g = genRandomMCGP(30) res = mcMSTPrim(30, n.lambdas = 50) print(res$pareto.front)
```

Seems to be an example badly tagged in the source code in these lines:
https://github.com/jakobbossek/mcMST/blob/4211a328d05ef9fb54916bc3cac94876dba5df64/R/mcMSTPrim.R#L26-L28
