### 2017-08-22 jakobbossek/mcMST#10 @gvegayon
Specify (or actually point to) the definition of the output of mcMSTEmoaZhou and friends
The manual of the function mcMSTEmoaZhou does not specify what it returns.
