### 2017-08-22 jakobbossek/mcMST#11 @gvegayon
Coding standards/misc
I know that there is no golden rule/book of coding conventions in R. But I just thought about mentioning these that you can consider to add in your project (which overall looks very good :+1: ), plus a couple of other misc comments:

1. Following Hadley Wickham's R-pkgs

    > It’s common for packages to be listed in Imports in DESCRIPTION, but not in NAMESPACE. In fact, this is what I recommend: list the package in DESCRIPTION so that it’s installed, then always refer to it explicitly with pkg::fun(). Unless there is a strong reason not to, it’s better to be explicit. It’s a little more work to write, but a lot easier to read when you come back to the code in the future.

    http://r-pkgs.had.co.nz/namespace.html

    I would use the `::` notation instead of, for example, `requirePackages` in [plotGraph.R](https://github.com/jakobbossek/mcMST/blob/4211a328d05ef9fb54916bc3cac94876dba5df64/R/plotGraph.R#L22) I would directly call `ggplot2::ggplot` and friends.

2.  Since you already have the print method for `mcGP` objects, you can introduce the `plot.mcGP` method as well instead of the plotGraph function. Sort of the same thing applies with the objects of class `ecr_multi_objective_result` `ecr_result`

3. I miss some examples in: `edgeListToCharVec`, `permutationToCharVec`, `permutationToEdgelist`, `prueferToCharVec`

4.  Not an expert in the problem of multiple objective minimum spanning tree problems, but after googling a on CRAN I found these packages that are related and perhaps can be nice to include around references:
    *   https://cran.r-project.org/package=mopsocd (a bit old)
    *   https://cran.r-project.org/package=moko (a more recent problem)
    *   igraph has a MST function https://cran.r-project.org/package=igraph it may be good to point at it
    *   https://cran.r-project.org/web/packages/optrees/optrees.pdf which is old and depends on igraph
