### 2017-08-22 #374 @gvegayon
Overall the project looks very good, although a couple of things can be done to (I think) improve it by solving the following issues that I just listed:

1.  The AppVeyor link is broken (https://github.com/jakobbossek/mcMST/issues/8)
2.  The [@]return tag in mcMSTPrim.R is bad (https://github.com/jakobbossek/mcMST/issues/9)
3.  Specify (or actually point to) the definition of the output of mcMSTEmoaZhou and friends (https://github.com/jakobbossek/mcMST/issues/10)
4.  Coding standards/misc (https://github.com/jakobbossek/mcMST/issues/11)

HIH
