### 2017-08-22 #374 @jakobbossek
Hi guys,

thanks for editing @Kevin-Mattheus-Moerman and reviewing @gvegayon 👍
I'm on vacation right now forcing myself not to work nevertheless and read work-related mails too often 😉

@gvegayon Thanks for the fast review process! I am happy to go through the issues you opened as soon as vacation ends.
