### 2017-09-04 3714fe3
fix AppVeyor badge in README
