### 2017-09-04 fcd009b
fix missing @examples in docs of mcMSTPrim
