### 2017-09-04 1ab4ced
add related work section to README
