### 2017-09-04 b431c15
add more examples; improve docs of return value of EMOAs
