+++ man-roxygen/ret_charvec.R
@@ -1,1 +1,1 @@
-#' @return [integer] Characteristic vector cv with cv[i] = 1 if the i-th edge is in the tree.
+#' @return [\code{integer}] Characteristic vector cv with cv[i] = 1 if the i-th edge is in the tree.
