+++ man/prueferToEdgelist.Rd
@@ -16,6 +16,14 @@ Pruefer code encoding a minimum spanning tree.}
 \description{
 Convert Pruefer code to edge list.
 }
+\examples{
+# here we generate a random Pruefer-code representing
+# a random spanning tree of a graph with n = 10 nodes
+pcode = sample(1:10, 8, replace = TRUE)
+print(pcode)
+edgelist = prueferToEdgeList(pcode)
+print(edgelist)
+}
 \seealso{
 Other transformation functions: \code{\link{edgeListToCharVec}},
   \code{\link{permutationToCharVec}},
