### 2017-09-04 jakobbossek/mcMST#10 @jakobbossek
Fixed via b431c15
