### 2017-09-04 jakobbossek/mcMST#9 @jakobbossek
Added missing `@examples` tag 🙈
