### 2017-09-04 jakobbossek/mcMST#11 @jakobbossek
Thanks @gvegayon for all the valuable comments! 👏

ad 1) Thanks for pointing this out. Did not know about that recommendation yet. I would change this in all my projects later. Is this a blocker in your eyes?
ad 2) I get the point. However, `plot.mcGP` would expect to use base R graphics. I rather prefer `ggplot2` graphics. Yes, in this case I could provide `autoplot.mcGP`. However, this function is expected to return an object of type `gg`/`ggplot`, but plotGraph is composed of two graphics side by side in some cases, which would require to return a list of ggplot objects or a ggtable object. This is why I strongly prefer not to change this.
ad 3) added examples to all transformation functions.
ad 4) Thanks for the package list. Added a section on related work to the README.
