### 2017-09-05 d1c03b9
New plot.mcMG method
