+++ man/plot.mcGP.Rd
@@ -1,10 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/plotGraph.R
-\name{plotGraph}
-\alias{plotGraph}
+\name{plot.mcGP}
+\alias{plot.mcGP}
 \title{Visualize bi-objective graph.}
 \usage{
-plotGraph(graph, show.cluster.centers = TRUE, ...)
+\method{plot}{mcGP}(graph, show.cluster.centers = TRUE, x = NULL,
+  y = NULL, ...)
 }
 \arguments{
 \item{graph}{[\code{mcGP}]\cr
@@ -13,6 +14,10 @@ Multi-objective graph problem.}
 \item{show.cluster.centers}{[\code{logical(1)}]\cr
 Display cluster centers?
 Default is \code{TRUE}. This option is ignored silently if the instance is not clustered.}
+
+\item{x}{Not used at the moment.}
+
+\item{y}{Not used at the moment.}
 
 \item{...}{[any]\cr
 Not used at the moment.}
@@ -24,7 +29,7 @@ nodes). The latter is \code{NULL}, if \code{graph} has no associated coordinates
 }
 \description{
 Only applicable for bi-objective problems of class \code{mcGP}.
-\code{plotGraph} generates a scatterplot of edge weights. If the nodes do
+\code{plot.mcGP} generates a scatterplot of edge weights. If the nodes do
 have coordinates, additionally a scatterplot of the nodes in the euclidean
 plane is generated.
 }
