+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,mcGP)
 S3method(print,mcGP)
 export(addCenters)
 export(addCoordinates)
@@ -22,7 +23,6 @@ export(mutSubgraphMST)
 export(mutUniformPruefer)
 export(permutationToCharVec)
 export(permutationToEdgelist)
-export(plotGraph)
 export(prueferToCharVec)
 export(prueferToEdgeList)
 import(BBmisc)
