### 2017-09-05 jakobbossek/mcMST#11 @gvegayon
> Is this a blocker in your eyes?

Not at all, as I said, it is about style. I do think that HW's suggestion should be a standard overall, but if your package works and CRAN is OK with it, I should be OK as well.

> plot.mcGP would expect to use base R graphics

Why is that? I'm not a ggplot2 user, but I think there is no requirement for plot.[someclass]... it is actually very flexible. In the case of ggplot2 it suffices with returning the ggplot object and R will do the rest (no need of explicitly calling print). [Here](https://github.com/USCbiostats/aphylo/blob/1be4b478aa29598438bfbb9ef1ef0742903009b0/R/data-management.r#L440-L471) is an example of an R package that I've been working on that uses ggplot2 (actually ggtree) in a plot S3 method. Furthermore. I've just submitted a pull request with this so you can take a look at it: https://github.com/jakobbossek/mcMST/pull/12

> added examples to all transformation functions.

Cool!

> Thanks for the package list. Added a section on related work to the README.

You are welcomed

Also, I just realized that there is no manual for mcMST or mcMST-package. I suggest you create one... That would be #5

Also (2), I added a ChangeLog... I think [these are useful](https://en.wikipedia.org/wiki/Changelog)

HIH
