### 2017-09-05 6172690
Fixing plot.mcGP
