### 2017-09-07 9dfbeaf
add docs for package
