### 2017-09-07 cf97863
update paper: add reference to recently accepted SSCI paper
