+++ paper/paper.bib
@@ -8,13 +8,20 @@
 }
 
 @INPROCEEDINGS{KC01,
-  author={J. D. Knowles and D. W. Corne},
+  author={Knowles, J. D. and Corne, D. W.},
   booktitle={Proceedings of the 2001 Congress on Evolutionary Computation (IEEE Cat. No.01TH8546)},
   title={A comparison of encodings and algorithms for multiobjective minimum spanning tree problems},
   year={2001},
   volume={1},
   pages={544-551 vol. 1},
   doi={10.1109/CEC.2001.934439}
+}
+
+@INPROCEEDINGS{BG17,
+  author={Bossek, J. and Grimme, C.},
+  booktitle={{Proceedings of the 2017 IEEE Symposium Series on Computational Intelligence (accepted)}},
+  title={A Pareto-Beneficial Sub-Tree Mutation for the Multi-Criteria Minimum Spanning Tree Problem},
+  year={2017},
 }
 
 @article{P18,
@@ -27,7 +34,7 @@
 }
 
 @book{D01,
- author = {Deb, Kalyanmoy and Kalyanmoy, Deb},
+ author = {Deb, Kalyanmoy},
  title = {Multi-Objective Optimization Using Evolutionary Algorithms},
  year = {2001},
  isbn = {047187339X},
@@ -36,16 +43,16 @@
 }
 
 @article{ZG99,
-  author = "Gengui Zhou and Mitsuo Gen",
-  title = "Genetic algorithm approach on multi-criteria minimum spanning tree problem ",
-  journal = "European Journal of Operational Research ",
-  volume = "114",
-  number = "1",
-  pages = "141 - 152",
-  year = "1999",
-  issn = "0377-2217",
-  doi = "https://doi.org/10.1016/S0377-2217(98)00016-2",
-  url = "http://www.sciencedirect.com/science/article/pii/S0377221798000162",
+  author = {Zhou, Gengui and Gen, Mitsuo},
+  title = {Genetic algorithm approach on multi-criteria minimum spanning tree problem},
+  journal = {European Journal of Operational Research},
+  volume = {114},
+  number = {1},
+  pages = {141 - 152},
+  year = {1999},
+  issn = {0377-2217},
+  doi = {https://doi.org/10.1016/S0377-2217(98)00016-2},
+  url = {http://www.sciencedirect.com/science/article/pii/S0377221798000162},
 }
 
 @article {P57,
