### 2017-09-07 b23b89b
unify names in paper.bib
