+++ paper/paper.bib
@@ -34,7 +34,7 @@
 }
 
 @book{D01,
- author = {Deb, Kalyanmoy},
+ author = {Deb, K.},
  title = {Multi-Objective Optimization Using Evolutionary Algorithms},
  year = {2001},
  isbn = {047187339X},
@@ -43,7 +43,7 @@
 }
 
 @article{ZG99,
-  author = {Zhou, Gengui and Gen, Mitsuo},
+  author = {Zhou, G. and Gen, M.},
   title = {Genetic algorithm approach on multi-criteria minimum spanning tree problem},
   journal = {European Journal of Operational Research},
   volume = {114},
