### 2017-09-07 jakobbossek/mcMST#11 @jakobbossek
> > plot.mcGP would expect to use base R graphics
>
> Why is that? I'm not a ggplot2 user, but I think there is no requirement for plot.[someclass]... it is
> actually very flexible. In the case of ggplot2 it suffices with returning the ggplot object and R will do
> the rest (no need of explicitly calling print). Here is an example of an R package that I've been
> working on that uses ggplot2 (actually ggtree) in a plot S3 method. Furthermore. I've just submitted
> a pull request with this so you can take a look at it: #12

Thanks a lot @gvegayon! Merged.

> Also, I just realized that there is no manual for mcMST or mcMST-package. I suggest you create
> one... That would be #5

Just added docs for the package via 9dfbeaf 👍

> Also (2), I added a ChangeLog... I think these are useful

Thanks!

I guess everything is fine now 😃
