### 2017-09-07 #374 @gvegayon
Just closed the last issue created in jakobbossek/mcMST (https://github.com/jakobbossek/mcMST/issues/11). I'm OK with what the project looks like now :+1:
