### 2017-09-12 #374 @Kevin-Mattheus-Moerman
Thanks for the review @gvegayon! @arfon, over to you, we are good to accept this one!
