+++ NEWS.md
@@ -1,3 +1,8 @@
+# mcMST 1.0.1
+
+* improved documentation
+* plotGraph replaced by S3 method plot.mcGP
+
 # mcMST 1.0.0
 
 * first release to CRAN
