### 2017-09-18 #374 @jakobbossek
@gvegayon Thanks for the thorough review! 👍

@arfon Just uploaded a new version to [CRAN](https://cran.r-project.org/web/packages/mcMST/index.html) and [Zenedo](https://zenodo.org/record/894402#.Wb_Q1K3qi34). The DOI is https://doi.org/10.5281/zenodo.894402
